Extract icon settings to its own page
