+++ overwolf/src/AppSettings/AppSettings.tsx
@@ -1,13 +1,10 @@
 import clsx from 'clsx';
-import produce from 'immer';
 import React, { useContext, useState } from 'react';
-import { AppContext, AppContextSettings } from '@/contexts/AppContext';
+import { AppContext, AppContextSettings, IAppContext } from '@/contexts/AppContext';
 import ReturnIcon from '@/Icons/ReturnIcon';
-import SelectIcon from '@/Icons/SelectIcon';
-import UnselectIcon from '@/Icons/UnselectIcon';
-import { SimpleStorageSetting, store, storeIconCategory, storeIconType } from '@/logic/storage';
-import { compareNames } from '@/logic/util';
+import { SimpleStorageSetting, store } from '@/logic/storage';
 import { useAppSettingsStyles } from './appSettingsStyle';
+import IconSettingsPage from './pages/IconSettingsPage';
 import OverlaySettingsPage from './pages/OverlaySettingsPage';
 import WindowSettingsPage from './pages/WindowSettingsPage';
 
@@ -19,6 +16,7 @@ interface IProps {
 export interface IAppSettingsPageProps {
     settings: AppContextSettings;
     updateSimpleSetting: <TKey extends keyof SimpleStorageSetting>(key: TKey, value: SimpleStorageSetting[TKey]) => void;
+    updateSettings: IAppContext['update']
     setPeek: (peek: boolean) => void;
 }
 
@@ -32,90 +30,9 @@ export default function AppSettings(props: IProps) {
 
     const [isPeeking, setIsPeeking] = useState(false);
 
-    function updateIconCategorySettings(name: string, value: boolean) {
-        const settings = context.settings.iconSettings;
-        storeIconCategory(name, value);
-        if (settings) {
-            return produce(settings, draft => {
-                draft.categories[name].value = value;
-            });
-        }
-        return settings;
-    }
-
     function updateSimpleSetting<TKey extends keyof SimpleStorageSetting>(key: TKey, value: SimpleStorageSetting[TKey]) {
         store(key, value);
         context.update({ [key]: value });
-    }
-
-    function updateIconSettings(category: string, type: string, value: boolean) {
-        const settings = context.settings.iconSettings;
-        storeIconType(category, type, value);
-        if (settings) {
-            return produce(settings, draft => {
-                draft.categories[category].types[type].value = value;
-            });
-        }
-        return settings;
-    }
-
-    function selectAllIconsByCategory(category: string, value: boolean) {
-        const settings = context.settings.iconSettings;
-        if (settings) {
-            return produce(settings, draft => {
-                storeIconCategory(category, value);
-                draft.categories[category].value = value;
-                for (const type in draft.categories[category].types) {
-                    draft.categories[category].types[type].value = value;
-                    storeIconType(category, type, value);
-                }
-            });
-        }
-        return settings;
-    }
-
-    function renderIconFilterSettings() {
-        if (!context.settings.iconSettings) {
-            return null;
-        }
-
-        return Object.entries(context.settings.iconSettings.categories).sort(compareNames).map(([categoryKey, category]) => {
-            const typeChildren = Object.entries(category.types).sort(compareNames).map(([typeKey, type]) => {
-                return <p key={'FrameMenuType' + typeKey}>
-                    <label className={classes.checkbox}>
-                        <input
-                            type='checkbox'
-                            checked={type.value}
-                            onChange={e => context.update({ iconSettings: updateIconSettings(categoryKey, typeKey, e.currentTarget.checked) })}
-                        />
-                        {type.name}
-                    </label>
-                </p>;
-            });
-
-            return <details key={'FrameMenuCat' + categoryKey}>
-                <summary className={clsx(classes.summary, classes.iconCategory)}>
-                    <label className={classes.checkbox}>
-                        <input
-                            type='checkbox'
-                            checked={category.value}
-                            onChange={e => context.update({ iconSettings: updateIconCategorySettings(categoryKey, e.currentTarget.checked) })}
-                        />
-                        {category.name}
-                    </label>
-                    <span />
-                    <button className={classes.selectIcon} onClick={() => context.update({ iconSettings: selectAllIconsByCategory(categoryKey, true) })}>
-                        <SelectIcon />
-                    </button>
-                    <button className={classes.selectIcon} onClick={() => context.update({ iconSettings: selectAllIconsByCategory(categoryKey, false) })}>
-                        <UnselectIcon />
-                    </button>
-                </summary>
-                <div className={classes.iconTypeContainer}>
-                    {typeChildren}
-                </div>
-            </details>;
-        });
     }
 
     const rootClassName = clsx(
@@ -123,6 +40,13 @@ export default function AppSettings(props: IProps) {
         !visible && classes.hidden,
         context.settings.showHeader && classes.belowHeader,
         isPeeking && context.gameRunning && classes.peek);
+
+    const pageProps: IAppSettingsPageProps = {
+        settings: context.settings,
+        updateSimpleSetting,
+        updateSettings: context.update,
+        setPeek: setIsPeeking,
+    };
 
     return <div className={rootClassName}>
         <button className={classes.return} onClick={onClose}>
@@ -134,27 +58,19 @@ export default function AppSettings(props: IProps) {
             <details>
                 <summary className={classes.summary}>This window</summary>
                 <div className={classes.indent}>
-                    <WindowSettingsPage
-                        settings={context.settings}
-                        updateSimpleSetting={updateSimpleSetting}
-                        setPeek={setIsPeeking}
-                    />
+                    <WindowSettingsPage {...pageProps} />
                 </div>
             </details>
             <details>
                 <summary className={classes.summary}>In-game overlay window</summary>
                 <div className={classes.indent}>
-                    <OverlaySettingsPage
-                        settings={context.settings}
-                        updateSimpleSetting={updateSimpleSetting}
-                        setPeek={setIsPeeking}
-                    />
+                    <OverlaySettingsPage {...pageProps} />
                 </div>
             </details>
             <details>
                 <summary className={classes.summary}>Icon Categories</summary>
                 <div className={classes.indent}>
-                    {renderIconFilterSettings()}
+                    <IconSettingsPage {...pageProps} />
                 </div>
             </details>
         </div>
